Add lab token id command (#693)
* Add lab token id command

* fix doc

* Keep only the strkey id

* refactoring

---------
